feat: vendir sync caching
On vendir sync a new vendir.sync.yaml is safed to the "tmp" folder containing the current state of vendir.yaml condensed into hash values. If this file remains unchanged, vendir sync can be omitted.

The feature can be enabled on environment and application level with the new application.cache.enabled flag.