test(render): add integration tests to check rendering output (#118)
All folders in example folder are considered a myks rendered repository.
The test in internal/integration/ will rerender all envs and
applications. Nothing should be changed afterwards.

This enforces that every PR will contain all the introduced changes in
the rendered output (breaking change) as part of the PR.

This PR is huge due to the examples folders. The examples should be save
to be ignored for now.

---------

Co-authored-by: German Lashevich <german.lashevich@gmail.com>