fix: ensure buildDependencies is nullable
In YTT an annotation affects either its left node or its right node, in this
exact order. To make sure the annotation is applied to the bottom node, we have
to ensure there's no node to the left of it.

In this case, the `releaseName` key is moved to the top of the list, to achieve
that.